bug fix by turning E to float matrix

+++ interface.cpp
@@ -66,8 +66,8 @@ int main(int argc,char*argv[]) {
    std::vector<double> vec_score;
     
   //In order to illustrate the gvld(or vld)-consistant neighbors, the following two parameters has been externalized as inputs of the function KVLD.
-    libNumerics::matrix<int> E = libNumerics::matrix<int>::ones(matches.size(),matches.size())*(-1);
-    // gvld-consistancy matrix, intitialized to -1,  >0, -1=unknow, -2=false  
+    libNumerics::matrix<float> E = libNumerics::matrix<float>::ones(matches.size(),matches.size())*(-1);
+    // gvld-consistancy matrix, intitialized to -1,  >0 consistancy value, -1=unknow, -2=false  
     std::vector<bool> valide(matches.size(), true);// indices of match in the initial matches, if true at the end of KVLD, a match is kept.
 
     size_t it_num=0;
